[refactoring; avoid memory leaks]
