+++ src/ffi/boxer.rs
@@ -1,0 +1,10 @@
+use std::mem;
+
+pub fn boxed<T>(obj: T) -> *mut T { Box::into_raw(Box::new(obj)) }
+
+pub fn boxed_vec<T>(vec: Vec<T>) -> *mut T {
+    let mut slice = vec.into_boxed_slice();
+    let ptr = slice.as_mut_ptr();
+    mem::forget(slice);
+    ptr
+}
